refactor (SC-835): build_uuid_scale_factor_dict method to pm_helper
